photogram/requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-channels==0.7.0
django-cors-headers==3.13.0
django-filter==22.1
django-location-field==2.1.0
djangorestframework==3.14.0
drf-extra-fields==3.4.1
idna==3.4
Markdown==3.4.1
oauthlib==3.2.2
Pillow==9.3.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.3
urllib3==1.26.13
